[#44] Refactor: 조회 서비스 수정
