+++ src/main/java/umc/GrowIT/Server/service/ChallengeService/ChallengeQueryService.java
@@ -11,4 +11,5 @@ public interface ChallengeQueryService {
     String getUserDate(Long userId);
     ChallengeResponseDTO.ChallengeHomeDTO getChallengeHome(Long userId);
     ChallengeResponseDTO.ChallengeStatusListDTO getChallengeStatus(Long userId, ChallengeType dtype, Boolean completed);
+    ChallengeResponseDTO.ProofDetailsDTO getChallengeProofDetails(Long userId, Long challengeId);
 }
